+++ src/YosemitechModbus.cpp
@@ -47,6 +47,7 @@ String yosemitech::getModel(void)
         case Y550: {return "Y550";}
         case Y551: {return "Y551";}
         case Y560: {return "Y560";}
+        case Y700: {return "Y700";}
         case Y4000: {return "Y4000";}
         default:  {return "Unknown";}
     }
@@ -71,6 +72,7 @@ String yosemitech::getParameter(void)
         case Y550: {return "COD";}
         case Y551: {return "COD";}
         case Y560: {return "Ammonium";}
+        case Y700: {return "Pressure";}
         case Y4000: {return "DO,   Turb, Cond,  pH,   Temp, ORP,  Chl,  BGA";}
         default:  {return "Unknown";}
     }
@@ -95,6 +97,7 @@ String yosemitech::getUnits(void)
         case Y550: {return "mg/L, NTU";}
         case Y551: {return "mg/L, NTU";}
         case Y560: {return "mg/L";}
+        case Y700: {return "mm H2O";}        
         case Y4000: {return "mg/L, NTU,  mS/cm, pH,   °C,   mV,   µg/L, µg/L";}
         default:  {return "Unknown";}
     }
@@ -109,8 +112,9 @@ byte yosemitech::getSlaveID(void)
     byte command[8] = {0xFF, 0x03, 0x30, 0x00, 0x00, 0x01, 0x9E, 0xD4};
     int respSize = modbus.sendCommand(command, 8);
 
-    if (respSize == 7) return modbus.responseBuffer[3];
-    else return 0x01;  // This is the default address
+    // if (respSize == 7) return modbus.responseBuffer[3];
+    // else return 0x01;  // This is the default address
+    return modbus.responseBuffer[3];
 }
 
 
@@ -136,8 +140,8 @@ String yosemitech::getSerialNumber(void)
             SN = modbus.StringFromRegister(0x03, 0x0900, 14); break; // for all sensors except Y4000
     }
 
-    // Strip out the initial ')' that seems to come with some responses
-    if (SN.startsWith(")"))
+    // Strip out the initial ')' or '$' that seems to come with some responses
+    if (SN.startsWith(")") || SN.startsWith("$"))
     {
         SN = SN.substring(1);
     }
@@ -154,10 +158,12 @@ String yosemitech::getSerialNumber(void)
         if (modelSS == 9) _model = Y520;  // 09 means conductivity sensor
         if (modelSS == 10) _model = Y510;  // 10 means turbidity sensor
         if (modelSS == 29) _model = Y511;  // 29 means self-cleaning turbidity sensor
+        if (modelSS == 61) _model = Y513;  // 61 means Blue Green Algae (BGA)
         if (modelSS == 48) _model = Y514;  // 48 means chlorophyll
         if (modelSS == 43) _model = Y532;  // 43 must mean pH
         if (modelSS == 47) _model = Y551;  // 47 must mean COD
         if (modelSS == 68) _model = Y560;  // 68 must mean Ammonium
+        if (modelSS == 24) _model = Y700;  // 24 must mean Pressure/Depth
         if (modelSS == 38) _model = Y4000;  // 38 must mean MultiParameter Sonde
     }
 
@@ -213,12 +219,14 @@ bool yosemitech::startMeasurement(void)
             if (respSize == 8 && modbus.responseBuffer[0] == _slaveID) return true;
             else return false;
         }
-        // Y532 (pH), Y533 (ORP), Y560 (Ammonium) ion selective elctrodes do not
-        // require Start/Stop functions, which are not listed in their Modbus Manuals.
+        // Y532 (pH), Y533 (ORP), Y560 (Ammonium) ion selective elctrodes, and
+        // Y700 (Pressure/Depth) sensors do not require Start/Stop functions. 
+        // These commands are not in their Modbus Manuals.
         // However, Start/Stop functions are required to get these to work in ModularSensors.
         case Y532:
         case Y533:
         case Y560:
+        case Y700:
         {
             return true;
         }
@@ -243,12 +251,14 @@ bool yosemitech::stopMeasurement(void)
 {
     switch (_model)
     {
-        // Y532 (pH), Y533 (ORP), Y560 (Ammonium) ion selective elctrodes do not
-        // require Start/Stop functions, which are not listed in their Modbus Manuals.
+        // Y532 (pH), Y533 (ORP), Y560 (Ammonium) ion selective elctrodes, and
+        // Y700 (Pressure/Depth) sensors do not require Start/Stop functions. 
+        // These commands are not in their Modbus Manuals.
         // However, Start/Stop functions are required to get these to work in ModularSensors.
         case Y532:
         case Y533:
         case Y560:
+        case Y700:
         {
             return true;
         }
@@ -438,6 +448,19 @@ bool yosemitech::getValues(float &parmValue, float &tempValue, float &thirdValue
             }
             break;
         }
+        // Y700 Pressure/Depth
+        case Y700:
+        {
+            if (modbus.getRegisters(0x03, 0x2600, 6))
+            {
+                parmValue = modbus.float32FromFrame(littleEndian, 7);
+                errorCode = modbus.byteFromFrame(11);
+                // Get temperature at register 0x2400, after Frame is read
+                tempValue = modbus.float32FromRegister(0x03,  0x2400, littleEndian);
+                return true;
+            }
+            break;
+        }
         // Everybody else other than Y4000 Sonde; Y551 COD; Y532 (pH); Y533 (ORP); Y502 & Y504 (DO)
         default:
         {
@@ -481,7 +504,9 @@ bool yosemitech::getValues(float &parmValue)
     return getValues(parmValue, errorCode);
 }
 
-// Get 8 values for the multiparameter sonde, with or without error flag
+// Get 8 values for the Y4000 multiparameter sonde, with or without error flag
+// Note that only 6 sensors can be connected at a time,
+// so only 7 values (including temperature) will be returned.
 bool yosemitech::getValues(float &DOmgL, float &Turbidity, float &Cond,
                            float &pH, float &Temp, float &ORP,
                            float &Chlorophyll, float &BGA, byte &errorCode)
@@ -502,7 +527,7 @@ bool yosemitech::getValues(float &DOmgL, float &Turbidity, float &Cond,
         case Y4000:   // Y4000 Multiparameter sonde
         {
             // Sonde's 8 values begin in register 260
-            if (modbus.getRegisters(0x03, 0x2601, 10))
+            if (modbus.getRegisters(0x03, 0x2601, 16))  // Modbus manual has error!
             {
                 DOmgL   = modbus.float32FromFrame(littleEndian, 3);   // DOmgL
                 Turbidity = modbus.float32FromFrame(littleEndian, 7);   // Turbidity